+++ src/js/player.js
@@ -1084,7 +1084,6 @@ vjs.Player.prototype.exitFullWindow = function(){
 };
 
 vjs.Player.prototype.selectSource = function(sources){
-
   // Loop through each playback technology in the options order
   for (var i=0,j=this.options_['techOrder'];i<j.length;i++) {
     var techName = vjs.capitalize(j[i]),
@@ -1173,7 +1172,14 @@ vjs.Player.prototype.src = function(source){
 
       // wait until the tech is ready to set the source
       this.ready(function(){
-        this.techCall('src', source.src);
+
+        // The setSource tech method was added with source handlers
+        // so older techs won't support it
+        if (this.tech['setSource']) {
+          this.techCall('setSource', source);
+        } else {
+          this.techCall('src', source.src);
+        }
 
         if (this.options_['preload'] == 'auto') {
           this.load();